requirements.txt
python-telegram-bot==13.15
pyrebase4==4.5.0
requests==2.31.0
urllib3==2.2.1
certifi>=2023.7.22
idna>=3.4
chardet>=5.1.0
